backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-multipart
sentence-transformers
openai
chromadb
spacy
PyPDF2